Check assets:precompile and assets:clobber are defined before enhancing (#70)
* Check assets:precompile is defined before enhancing

As "Sprockets is not a dependency", we should not
assume that the assets:precompile task will be defined. If it's not
there, we should just skip enhancing it.

* Check assets:clobber is defined before enhancing

As "Sprockets is not a dependency", we should not
assume that the assets:precompile AND assets:clobber tasks will be defined. If it's not
there, we should just skip enhancing it.